Try to improve the adapters
