+++ modules/app/src/main/scala/com/fortysevendeg/ninecardslauncher/app/ui/commons/adapters/apps/AppsAdapter.scala
@@ -2,17 +2,16 @@ package com.fortysevendeg.ninecardslauncher.app.ui.commons.adapters.apps
 
 import android.support.v7.widget.GridLayoutManager
 import android.support.v7.widget.GridLayoutManager.SpanSizeLookup
-import android.view.View.{OnLongClickListener, OnClickListener}
-import android.view.{View, LayoutInflater, ViewGroup}
+import android.view.{LayoutInflater, ViewGroup}
 import com.fortysevendeg.macroid.extras.ResourcesExtras._
 import com.fortysevendeg.ninecardslauncher.app.ui.collections.actions.apps.ViewHolderAppLayoutAdapter
 import com.fortysevendeg.ninecardslauncher.app.ui.commons.Constants._
 import com.fortysevendeg.ninecardslauncher.app.ui.commons.UiContext
 import com.fortysevendeg.ninecardslauncher.app.ui.commons.adapters.{ScrollableManager, HeaderedItemAdapter, ItemHeaderedViewHolder}
 import com.fortysevendeg.ninecardslauncher.app.ui.commons.models.AppHeadered
-import com.fortysevendeg.ninecardslauncher.app.ui.components.FastScrollerListener
 import com.fortysevendeg.ninecardslauncher.process.device.models.App
-import com.fortysevendeg.ninecardslauncher2.R
+import com.fortysevendeg.ninecardslauncher2.{TR, R}
+import com.fortysevendeg.ninecardslauncher2.TypedResource._
 import macroid.ActivityContextWrapper
 
 case class AppsAdapter(
@@ -20,28 +19,15 @@ case class AppsAdapter(
   clickListener: (App) => Unit,
   longClickListener: Option[(App) => Unit])
   (implicit val activityContext: ActivityContextWrapper, implicit val uiContext: UiContext[_])
-  extends HeaderedItemAdapter[App]
-  with FastScrollerListener {
+  extends HeaderedItemAdapter[App] {
 
   val heightItem = resGetDimensionPixelSize(R.dimen.height_app_item)
 
-  override def createViewHolder(parent: ViewGroup): ItemHeaderedViewHolder[App] = {
-    val view = LayoutInflater.from(parent.getContext).inflate(R.layout.app_item, parent, false).asInstanceOf[ViewGroup]
-    view.setOnClickListener(new OnClickListener {
-      override def onClick(v: View): Unit = {
-        Option(v.getTag) foreach (tag => seq(Int.unbox(tag)).item foreach clickListener)
-      }
-    })
-    longClickListener foreach { listener =>
-      view.setOnLongClickListener(new OnLongClickListener {
-        override def onLongClick(v: View): Boolean = {
-          Option(v.getTag) foreach (tag => seq(Int.unbox(tag)).item foreach listener)
-          true
-        }
-      })
-    }
+  override def inflateView(parent: ViewGroup): ViewGroup =
+    LayoutInflater.from(parent.getContext).inflate(TR.layout.app_item, parent, false)
+
+  override def createViewHolder(view: ViewGroup): ItemHeaderedViewHolder[App] =
     new ViewHolderAppLayoutAdapter(view)
-  }
 
   override def getLayoutManager: GridLayoutManager = {
     val manager = new GridLayoutManager(activityContext.application, columnsLists) with ScrollableManager {
@@ -68,18 +54,4 @@ case class AppsAdapter(
     val heightItems = rowsWithItems._2 * heightItem
     heightHeaders + heightItems
   }
-
-  val defaultElement: Option[String] = None
-
-  override def getElement(position: Int): Option[String] = seq.foldLeft((defaultElement, false))((info, itemHeadered) =>
-    if (itemHeadered == seq(position)) {
-      (info._1, true)
-    } else {
-      (info._1, info._2) match {
-        case (_, false) => itemHeadered.header map (header => (Option(header), info._2)) getOrElse info
-        case _ => info
-      }
-    }
-  )._1
-
 }
